@@ -4,26 +4,27 @@ version: 0.0.2
 repository: https://github.com/ReniDelonzek/select_any
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
   intl: ^0.17.0
   data_table_plus: ^0.0.1
   flutter_date_pickers: ^0.2.4
-  share_extend: ^1.1.9
-  path_provider: ^2.0.1 
-  diacritic: ^0.1.1
-  bloc_pattern: ^2.5.1
-  mobx: ^1.1.0
-  flutter_mobx: ^1.1.0+2
-  msk_utils: ^0.0.3
+  share_extend: ^2.0.0
+  path_provider: ^2.0.2 
+  diacritic: ^0.1.3
+  bloc_pattern: ^3.0.0
+  mobx: ^2.0.4
+  flutter_mobx: ^2.0.2
+  msk_utils: ^0.0.4-dev.1
   flutter:
     sdk: flutter
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
-  build_runner: ^1.7.3
-  mobx_codegen: ^1.1.2
+  build_runner: ^2.1.1
+  mobx_codegen: ^2.0.3
   flutter_test:
     sdk: flutter
 
