@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
@@ -10,7 +11,7 @@ import 'package:select_any/src/modules/select_any_expanded/select_any_expanded_c
 class SelectAnyExpandedPage extends StatefulWidget {
   final SelectModel _selectModel;
   final ObservableList<ItemSelectExpanded> itens;
-  Map data;
+  Map? data;
 
   SelectAnyExpandedPage(this._selectModel, this.itens, {this.data});
 
@@ -21,14 +22,14 @@ class SelectAnyExpandedPage extends StatefulWidget {
 }
 
 class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
-  SelectAnyExpandedController controller;
+  late SelectAnyExpandedController controller;
   final GlobalKey<RefreshIndicatorState> _refreshIndicatorKey =
       new GlobalKey<RefreshIndicatorState>();
   bool loaded = false;
 
   // indica se está sendo usada a fonte alternativa ou nao
   bool fonteAlternativa = false;
-  BuildContext buildContext;
+  BuildContext? buildContext;
 
   _SelectAnyExpandedPageState(
       String title, ObservableList<ItemSelectExpanded> itens) {
@@ -56,7 +57,7 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
     return Scaffold(
       appBar: AppBar(
         centerTitle: true,
-        title: Observer(builder: (_) => controller.appBarTitle),
+        title: Observer(builder: (_) => controller.appBarTitle!),
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.endDocked,
       floatingActionButton: Padding(
@@ -85,7 +86,8 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
             itemCount: controller.listaExibida.length,
             itemBuilder: (context, index) {
               return Observer(
-                  builder: (_) => _getItemList(controller.listaExibida[index]));
+                  builder: (_) => _getItemList(
+                      controller.listaExibida[index] as ItemSelectExpanded));
             }),
       );
   }
@@ -110,7 +112,7 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
                         child: Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children:
-                          _getTexts(itemSelect.strings, itemSelect.object),
+                          _getTexts(itemSelect.strings!, itemSelect.object),
                     )),
                     if (itemSelect.items?.isNotEmpty == true)
                       IconButton(
@@ -124,7 +126,7 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
                 Observer(builder: (_) {
                   if (itemSelect.isExpanded) {
                     return Column(
-                        children: itemSelect.items
+                        children: itemSelect.items!
                             .map((element) => _getItemList(element))
                             .toList());
                   } else {
@@ -137,34 +139,34 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
     );
   }
 
-  Widget _getLinha(MapEntry item, Map map) {
-    Line linha = widget._selectModel.lines
-        .firstWhere((linha) => linha.key == item.key, orElse: () => null);
+  Widget? _getLinha(MapEntry item, Map? map) {
+    Line? linha = widget._selectModel.lines
+        .firstWhereOrNull((linha) => linha.key == item.key);
     if (linha == null) {
       return null;
     }
     dynamic valor = (item.value == null || item.value.toString().isEmpty)
-        ? (linha.defaultValue != null ? linha.defaultValue(map) : '')
+        ? (linha.defaultValue != null ? linha.defaultValue!(map) : '')
         : item.value;
     if (linha.formatData != null) {
-      valor = linha.formatData.formatData(ObjFormatData(data: valor, map: map));
+      valor =
+          linha.formatData!.formatData(ObjFormatData(data: valor, map: map));
     }
-    if (linha != null &&
-        (linha.enclosure != null || linha.customLine != null)) {
+    if ((linha.enclosure != null || linha.customLine != null)) {
       if (linha.customLine != null) {
-        return linha.customLine(CustomLineData(data: map));
+        return linha.customLine!(CustomLineData(data: map));
       }
-      return Text(linha.enclosure.replaceAll('???', valor),
+      return Text(linha.enclosure!.replaceAll('???', valor),
           style: linha.textStyle);
     } else {
-      return Text(valor?.toString(), style: linha.textStyle);
+      return Text(valor.toString(), style: linha.textStyle);
     }
   }
 
-  List<Widget> _getTexts(Map<String, dynamic> map, Map object) {
+  List<Widget> _getTexts(Map<String, dynamic> map, Map? object) {
     List<Widget> widgets = [];
     for (var item in map.entries) {
-      Widget widget = _getLinha(item, object);
+      Widget? widget = _getLinha(item, object);
       if (widget != null) {
         widgets.add(widget);
       }
@@ -172,12 +174,12 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
     return widgets;
   }
 
-  void _onAction(ItemSelect itemSelect, ActionSelect acao) async {
+  void _onAction(ItemSelect? itemSelect, ActionSelect acao) async {
     if (acao.function != null) {
       if (acao.closePage) {
         Navigator.pop(context);
       }
-      acao.function(DataFunction(data: itemSelect, context: context));
+      acao.function!(DataFunction(data: itemSelect, context: context));
     }
     if (acao.functionUpd != null) {
       if (acao.closePage) {
@@ -185,19 +187,19 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
       }
 
       var res = await acao
-          .functionUpd(DataFunction(data: itemSelect, context: context));
+          .functionUpd!(DataFunction(data: itemSelect, context: context));
       if (res == true) {
         loaded = false;
       }
     } else if (acao.route != null || acao.page != null) {
       Map<String, dynamic> dados = Map();
       if (acao.keys?.entries != null) {
-        for (MapEntry dado in acao.keys.entries) {
+        for (MapEntry dado in acao.keys!.entries) {
           if (itemSelect != null &&
               (itemSelect.object as Map).containsKey(dado.key)) {
             dados.addAll({dado.value: itemSelect.object[dado.key]});
-          } else if (widget.data.containsKey(dado.key)) {
-            dados.addAll({dado.value: widget.data[dado.key]});
+          } else if (widget.data!.containsKey(dado.key)) {
+            dados.addAll({dado.value: widget.data![dado.key]});
           }
         }
       }
@@ -211,9 +213,9 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
           : RouteSettings();
 
       var res = await Navigator.of(context).push(acao.route != null
-          ? acao.route
+          ? acao.route!
           : new MaterialPageRoute(
-              builder: (_) => acao.page, settings: settings));
+              builder: (_) => acao.page!, settings: settings));
       if (acao.closePage) {
         if (res != null) {
           if (res is Map &&
@@ -254,7 +256,7 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
     //       child: Icon(Icons.filter_list)));
     // }
     if (!(widget._selectModel.buttons?.isEmpty ?? true)) {
-      for (ActionSelect acao in widget._selectModel.buttons) {
+      for (ActionSelect acao in widget._selectModel.buttons!) {
         widgets.add(FloatingActionButton(
           heroTag: widgets.length,
           mini: widgets.isNotEmpty,
@@ -276,9 +278,9 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
         builder: (BuildContext bc) {
           return Container(
             child: new Wrap(
-              children: widget._selectModel.actions
+              children: widget._selectModel.actions!
                   .map((acao) => new ListTile(
-                      title: new Text(acao.description),
+                      title: new Text(acao.description!),
                       onTap: () {
                         Navigator.pop(context);
                         _onAction(itemSelect, acao);
@@ -292,10 +294,10 @@ class _SelectAnyExpandedPageState extends State<SelectAnyExpandedPage> {
   void _tratarOnTap(ItemSelect itemSelect) {
     if (widget._selectModel.typeSelect == TypeSelect.ACTION &&
         widget._selectModel.actions != null) {
-      if (widget._selectModel.actions.length > 1) {
+      if (widget._selectModel.actions!.length > 1) {
         _exibirListaAcoes(itemSelect);
-      } else if (widget._selectModel.actions.isNotEmpty) {
-        ActionSelect acao = widget._selectModel.actions?.first;
+      } else if (widget._selectModel.actions!.isNotEmpty) {
+        ActionSelect? acao = widget._selectModel.actions?.first;
         if (acao != null) {
           _onAction(itemSelect, acao);
         }
