@@ -11,7 +11,7 @@ abstract class _SelectFKBase with Store {
   @observable
   bool inFocus = false;
   @observable
-  Map<String, dynamic> obj;
+  Map<String, dynamic>? obj;
   @observable
   bool showClearIcon = false;
 
@@ -20,14 +20,14 @@ abstract class _SelectFKBase with Store {
   @observable
   bool listIsLoaded = false;
   FocusNode focusNode = FocusNode();
-  SelectModel selectModel;
+  SelectModel? selectModel;
 
   /// Retorna o valor da chave, caso o objeto não seja null e o valor conste no objeto
   getValueKey(String key) {
-    if (obj == null || !obj.containsKey(key)) {
+    if (obj == null || !obj!.containsKey(key)) {
       return null;
     }
-    return obj[key];
+    return obj![key];
   }
 
   /// Verifica se a [fontData] especificada retorna somente um registro
@@ -35,9 +35,9 @@ abstract class _SelectFKBase with Store {
   void checkSingleRow() async {
     /// Deixa o limite como dois, porque caso retorne dois ele possui > 1 registro
     /// Pode ser null caso o widget não tenha sido construído ainda
-    selectModel?.dataSource?.getList(2, 0, selectModel)?.then((value) {
+    selectModel?.dataSource.getList(2, 0, selectModel).then((value) {
       value.first.then((value) {
-        if (value.data?.length == 1) {
+        if (value.data.length == 1) {
           obj = value.data.first.object;
         }
       });
@@ -60,7 +60,7 @@ abstract class _SelectFKBase with Store {
   carregarDados() {
     if (!listIsLoaded) {
       /// Pode ser null caso o widget não tenha sido construído ainda
-      selectModel?.dataSource?.getList(-1, -1, selectModel)?.then((value) {
+      selectModel?.dataSource.getList(-1, -1, selectModel).then((value) {
         value.listen((event) {
           listIsLoaded = true;
           list = ObservableList.of(event.data);
