@@ -27,13 +27,13 @@ mixin _$SelectFKController on _SelectFKBase, Store {
   final _$objAtom = Atom(name: '_SelectFKBase.obj');
 
   @override
-  Map<String, dynamic> get obj {
+  Map<String, dynamic>? get obj {
     _$objAtom.reportRead();
     return super.obj;
   }
 
   @override
-  set obj(Map<String, dynamic> value) {
+  set obj(Map<String, dynamic>? value) {
     _$objAtom.reportWrite(value, super.obj, () {
       super.obj = value;
     });
